+++ src/enums.ts
@@ -1,11 +1,15 @@
 export enum TransactionType {
-  MINT,
-  SWAP,
-  REDEEM,
-  EXIT,
-  PAIDFEE,
-  SAVE,
-  WITHDRAW,
+  MASSET_MINT,
+  MASSET_SWAP,
+  MASSET_REDEEM,
+  MASSET_REDEEM_MASSET,
+  MASSET_PAID_FEE,
+  SAVINGS_CONTRACT_DEPOSIT,
+  SAVINGS_CONTRACT_WITHDRAW,
+  STAKING_REWARDS_CONTRACT_CLAIM_REWARD,
+  STAKING_REWARDS_CONTRACT_EXIT,
+  STAKING_REWARDS_CONTRACT_STAKE,
+  STAKING_REWARDS_CONTRACT_WITHDRAW,
 }
 
 export enum TimeMetricPeriod {
@@ -20,6 +24,40 @@ export enum TimeMetricPeriod {
 export enum AggregateMetricType {
   TOTAL_SUPPLY,
   TOTAL_SAVINGS,
+}
+
+export enum StakingRewardsContractType {
+  STAKING_REWARDS,
+  STAKING_REWARDS_WITH_PLATFORM_TOKEN,
+}
+
+export enum StakingRewardType {
+  REWARD,
+  PLATFORM_REWARD,
+}
+
+export function mapStakingRewardsContractType(
+  type: StakingRewardsContractType,
+): string {
+  switch (type) {
+    case StakingRewardsContractType.STAKING_REWARDS:
+      return 'STAKING_REWARDS'
+    case StakingRewardsContractType.STAKING_REWARDS_WITH_PLATFORM_TOKEN:
+      return 'STAKING_REWARDS_WITH_PLATFORM_TOKEN'
+    default:
+      return ''
+  }
+}
+
+export function mapStakingRewardType(type: StakingRewardType): string {
+  switch (type) {
+    case StakingRewardType.PLATFORM_REWARD:
+      return 'PLATFORM_REWARD'
+    case StakingRewardType.REWARD:
+      return 'REWARD'
+    default:
+      return ''
+  }
 }
 
 export function mapTimeMetricPeriod(period: TimeMetricPeriod): string {
@@ -43,20 +81,28 @@ export function mapTimeMetricPeriod(period: TimeMetricPeriod): string {
 
 export function mapTransactionType(type: TransactionType): string {
   switch (type) {
-    case TransactionType.MINT:
-      return 'MINT'
-    case TransactionType.SWAP:
-      return 'SWAP'
-    case TransactionType.REDEEM:
-      return 'REDEEM'
-    case TransactionType.EXIT:
-      return 'EXIT'
-    case TransactionType.PAIDFEE:
-      return 'PAIDFEE'
-    case TransactionType.SAVE:
-      return 'SAVE'
-    case TransactionType.WITHDRAW:
-      return 'WITHDRAW'
+    case TransactionType.MASSET_MINT:
+      return 'MASSET_MINT'
+    case TransactionType.MASSET_SWAP:
+      return 'MASSET_SWAP'
+    case TransactionType.MASSET_REDEEM:
+      return 'MASSET_REDEEM'
+    case TransactionType.MASSET_REDEEM_MASSET:
+      return 'MASSET_REDEEM_MASSET'
+    case TransactionType.MASSET_PAID_FEE:
+      return 'MASSET_PAID_FEE'
+    case TransactionType.SAVINGS_CONTRACT_DEPOSIT:
+      return 'SAVINGS_CONTRACT_DEPOSIT'
+    case TransactionType.SAVINGS_CONTRACT_WITHDRAW:
+      return 'SAVINGS_CONTRACT_WITHDRAW'
+    case TransactionType.STAKING_REWARDS_CONTRACT_CLAIM_REWARD:
+      return 'STAKING_REWARDS_CONTRACT_CLAIM_REWARD'
+    case TransactionType.STAKING_REWARDS_CONTRACT_EXIT:
+      return 'STAKING_REWARDS_CONTRACT_EXIT'
+    case TransactionType.STAKING_REWARDS_CONTRACT_STAKE:
+      return 'STAKING_REWARDS_CONTRACT_STAKE'
+    case TransactionType.STAKING_REWARDS_CONTRACT_WITHDRAW:
+      return 'STAKING_REWARDS_CONTRACT_WITHDRAW'
     default:
       return ''
   }
@@ -65,9 +111,9 @@ export function mapTransactionType(type: TransactionType): string {
 export function mapAggregateMetricType(type: AggregateMetricType): string {
   switch (type) {
     case AggregateMetricType.TOTAL_SAVINGS:
-      return 'TOTAL_SAVINGS';
+      return 'TOTAL_SAVINGS'
     case AggregateMetricType.TOTAL_SUPPLY:
-      return 'TOTAL_SUPPLY';
+      return 'TOTAL_SUPPLY'
     default:
       return ''
   }