+++ src/mappings/StakingRewards/StakingRewards.ts
@@ -1,0 +1,52 @@
+import {
+  RewardAdded,
+  RewardPaid,
+  Staked,
+  Withdrawn,
+} from '../../../generated/templates/StakingRewards/StakingRewards'
+import { StakingRewardsContractType } from '../../enums'
+import {
+  handleRewardAddedForType,
+  handleStakedForType,
+  handleRewardPaidForType,
+  handleWithdrawnForType,
+} from './shared'
+import {
+  getOrCreateStakingRewardsContractRewardPaidTransaction,
+  getOrCreateStakingRewardsContractWithdrawTransaction,
+  getOrCreateStakingRewardsContractStakeTransaction,
+} from '../../models/Transaction'
+
+export function handleRewardAdded(event: RewardAdded): void {
+  handleRewardAddedForType(
+    event.address,
+    StakingRewardsContractType.STAKING_REWARDS,
+  )
+}
+
+export function handleStaked(event: Staked): void {
+  handleStakedForType(
+    event.address,
+    StakingRewardsContractType.STAKING_REWARDS,
+    event.params.user,
+  )
+  getOrCreateStakingRewardsContractStakeTransaction(event)
+}
+
+export function handleWithdrawn(event: Withdrawn): void {
+  handleWithdrawnForType(
+    event.address,
+    StakingRewardsContractType.STAKING_REWARDS,
+    event.params.user,
+  )
+  getOrCreateStakingRewardsContractWithdrawTransaction(event)
+}
+
+export function handleRewardPaid(event: RewardPaid): void {
+  handleRewardPaidForType(
+    event.address,
+    StakingRewardsContractType.STAKING_REWARDS,
+    event.params.user,
+  )
+  getOrCreateStakingRewardsContractRewardPaidTransaction(event)
+}